requirements.txt
kagglehub
numpy
Pillow
pycocotools
torch
torchmetrics
torchvision
tqdm
ultralytics
transformers
timm

mmengine==0.10.7
mmdet==3.3.0
mmcv==2.1.0
nvidia-cuda-cccl==13.0.85
nvidia-cuda-crt==13.0.88
nvidia-cuda-nvcc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-nvvm==13.0.88
